requirements.txt
alembic==1.14.0
anyio==4.6.2.post1
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-telegram-bot==21.7
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.6
WTForms==3.2.1
zipp==3.20.2
